task(SDK-4214) - Removes logic from enum class and adds to a function
